refactor(scram): rename Keystore credential store to ScramCredentialFile
Align code terminology with documentation, which consistently refers
to the artifact as a "proxy SCRAM credential file". Renames the module,
package, and all Keystore-prefixed classes. Updates cross-module
references in POM files, documentation, integration tests, and the
CLI launcher.
