+++ kroxylicious-app/pom.xml
@@ -409,7 +409,7 @@
                 </dependency>
                 <dependency>
                     <groupId>io.kroxylicious</groupId>
-                    <artifactId>kroxylicious-scram-credential-store-provider-keystore</artifactId>
+                    <artifactId>kroxylicious-scram-credential-store-provider-file</artifactId>
                     <scope>runtime</scope>
                 </dependency>
                 <dependency>
