+++ kroxylicious-bom/pom.xml
@@ -186,7 +186,7 @@
 
             <dependency>
                 <groupId>io.kroxylicious</groupId>
-                <artifactId>kroxylicious-scram-credential-store-provider-keystore</artifactId>
+                <artifactId>kroxylicious-scram-credential-store-provider-file</artifactId>
                 <version>${project.version}</version>
             </dependency>
