+++ kroxylicious-docs/docs/_modules/sasl-termination/con-example-proxy-config.adoc
@@ -30,7 +30,7 @@ filterDefinitions:
           expectedAudience: kafka
           expectedIssuer: https://idp.example.com
         - mechanism: SCRAM-SHA-256
-          credentialStore: KeystoreScramCredentialStoreService
+          credentialStore: ScramCredentialFileService
           credentialStoreConfig:
             file: /etc/kroxylicious/scram-credentials.p12
             storePassword:
@@ -77,13 +77,13 @@ The following properties are optional:
 The `SCRAM-SHA-256` and `SCRAM-SHA-512` mechanisms validate client credentials against a pluggable credential store.
 
 * `credentialStore` (required) specifies the name of a plugin implementing `ScramCredentialStoreService`.
-The built-in KeyStore credential store is named `KeystoreScramCredentialStoreService`.
+The built-in credential file store is named `ScramCredentialFileService`.
 * `credentialStoreConfig` (required) provides configuration for the credential store plugin.
 The structure depends on the credential store implementation.
 
-=== KeyStore credential store configuration
+=== Credential file store configuration
 
-The `KeystoreScramCredentialStoreService` credential store loads SCRAM credentials from a proxy SCRAM credential file.
+The `ScramCredentialFileService` credential store loads SCRAM credentials from a proxy SCRAM credential file.
 Use the SCRAM credential tool to create and populate this file.
 See <<proc-managing-scram-credentials-{context}>> for instructions.
