+++ kroxylicious-filters/kroxylicious-sasl-termination/src/main/java/io/kroxylicious/filter/sasl/termination/ScramMechanismConfig.java
@@ -37,8 +37,8 @@ public record ScramMechanismConfig(
 
     public static final String MECHANISM_NAME_SCRAM_SHA_256 = "SCRAM-SHA-256";
     public static final String MECHANISM_NAME_SCRAM_SHA_512 = "SCRAM-SHA-512";
-    // Matches KeystoreCredentialManager.DEFAULT_ITERATIONS so phantom challenges are
-    // indistinguishable from real ones when using the default keystore credential store.
+    // Matches ScramCredentialFileManager.DEFAULT_ITERATIONS so phantom challenges are
+    // indistinguishable from real ones when using the default credential file store.
     static final int DEFAULT_PHANTOM_ITERATIONS = 10_000;
 
     /** Validates that credential store settings are present. */
