+++ kroxylicious-integration-tests/pom.xml
@@ -76,7 +76,7 @@
         </dependency>
         <dependency>
             <groupId>io.kroxylicious</groupId>
-            <artifactId>kroxylicious-scram-credential-store-provider-keystore</artifactId>
+            <artifactId>kroxylicious-scram-credential-store-provider-file</artifactId>
             <scope>test</scope>
         </dependency>
         <dependency>
