+++ kroxylicious-scram-credential-store-providers/kroxylicious-scram-credential-store-provider-file/src/main/resources/META-INF/services/io.kroxylicious.scram.credentialstore.ScramCredentialStoreService
@@ -1,0 +1,1 @@
+io.kroxylicious.scram.credentialstore.file.ScramCredentialFileService
