+++ kroxylicious-scram-credential-store-providers/kroxylicious-scram-credential-store-provider-file/src/test/java/io/kroxylicious/scram/credentialstore/file/ScramCredentialFileServiceTest.java
@@ -1,0 +1,147 @@
+/*
+ * Copyright Kroxylicious Authors.
+ *
+ * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
+ */
+
+package io.kroxylicious.scram.credentialstore.file;
+
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.attribute.PosixFileAttributeView;
+import java.nio.file.attribute.PosixFilePermissions;
+
+import org.apache.kafka.common.security.scram.internals.ScramMechanism;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.io.TempDir;
+
+import io.kroxylicious.proxy.config.secret.InlinePassword;
+import io.kroxylicious.scram.credentialstore.ScramCredentialStore;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatCode;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+
+class ScramCredentialFileServiceTest {
+
+    private static final String STORE_PASSWORD = "test-password";
+
+    @TempDir
+    Path tempDir;
+
+    private Path keystorePath;
+    private ScramCredentialFileConfig config;
+    private ScramCredentialFileService service;
+
+    @BeforeEach
+    void setUp() throws Exception {
+        keystorePath = tempDir.resolve("test-credentials.p12");
+
+        var generator = new TestCredentialGenerator();
+        generator.generateKeyStore(keystorePath, STORE_PASSWORD, ScramMechanism.SCRAM_SHA_256, "alice", "alice-secret");
+
+        if (Files.getFileAttributeView(keystorePath, PosixFileAttributeView.class) != null) {
+            Files.setPosixFilePermissions(keystorePath, PosixFilePermissions.fromString("rw-------"));
+        }
+
+        config = new ScramCredentialFileConfig(
+                keystorePath.toString(),
+                new InlinePassword(STORE_PASSWORD));
+
+        service = new ScramCredentialFileService();
+    }
+
+    @AfterEach
+    void tearDown() throws Exception {
+        if (service != null) {
+            service.close();
+        }
+        if (keystorePath != null && Files.exists(keystorePath)) {
+            Files.delete(keystorePath);
+        }
+    }
+
+    @Test
+    void shouldInitializeSuccessfully() {
+        assertThatCode(() -> service.initialize(config))
+                .doesNotThrowAnyException();
+    }
+
+    @Test
+    void shouldBuildCredentialStoreAfterInitialize() {
+        service.initialize(config);
+        ScramCredentialStore store = service.buildCredentialStore();
+
+        assertThat(store).isNotNull()
+                .isInstanceOf(ScramCredentialFile.class);
+    }
+
+    @Test
+    void shouldBuildMultipleStores() {
+        service.initialize(config);
+
+        ScramCredentialStore store1 = service.buildCredentialStore();
+        ScramCredentialStore store2 = service.buildCredentialStore();
+
+        assertThat(store1).isNotNull();
+        assertThat(store2).isNotNull();
+        assertThat(store1).isNotSameAs(store2);
+    }
+
+    @Test
+    void shouldRejectBuildBeforeInitialize() {
+        assertThatThrownBy(() -> service.buildCredentialStore())
+                .isInstanceOf(IllegalStateException.class)
+                .hasMessageContaining("not been initialized");
+    }
+
+    @Test
+    void shouldRejectDoubleInitialize() {
+        service.initialize(config);
+
+        assertThatThrownBy(() -> service.initialize(config))
+                .isInstanceOf(IllegalStateException.class)
+                .hasMessageContaining("already been initialized");
+    }
+
+    @Test
+    void shouldRejectBuildAfterClose() {
+        service.initialize(config);
+        service.close();
+
+        assertThatThrownBy(() -> service.buildCredentialStore())
+                .isInstanceOf(IllegalStateException.class)
+                .hasMessageContaining("been closed");
+    }
+
+    @Test
+    void shouldRejectInitializeAfterClose() {
+        service.close();
+
+        assertThatThrownBy(() -> service.initialize(config))
+                .isInstanceOf(IllegalStateException.class)
+                .hasMessageContaining("been closed");
+    }
+
+    @Test
+    void shouldAllowMultipleClose() {
+        service.initialize(config);
+
+        assertThatCode(() -> {
+            service.close();
+            service.close();
+            service.close();
+        }).doesNotThrowAnyException();
+    }
+
+    @Test
+    void shouldAllowCloseWithoutInitialize() {
+        try (ScramCredentialFileService uninitializedService = new ScramCredentialFileService()) {
+
+            assertThatCode(uninitializedService::close)
+                    .doesNotThrowAnyException();
+        }
+    }
+}
