--- kroxylicious-scram-credential-store-providers/kroxylicious-scram-credential-store-provider-keystore/src/main/resources/META-INF/services/io.kroxylicious.scram.credentialstore.ScramCredentialStoreService
@@ -1,1 +1,0 @@
-io.kroxylicious.scram.credentialstore.keystore.KeystoreScramCredentialStoreService
